requirements.txt
xmltodict
openai
streamlit
python-dotenv